* Make navigation bar's space touchable when bars are hidden * Don't make photo view set itself as a delegate, so others can be the delegate * Make zoomEnabled actually work for scroll views * Optionally support a cancel button in search bars * Use the Search or Done return key for search fields depending on if search is automatic

+++ src/Three20/TTBackgroundView.h
@@ -8,7 +8,7 @@
   UIColor* _fillColor;
   UIColor* _fillColor2;
   UIColor* _strokeColor;
-  int _strokeRadius;
+  NSInteger _strokeRadius;
   UIEdgeInsets _backgroundInset;
 }
 
@@ -16,7 +16,7 @@
 @property(nonatomic,retain) UIColor* fillColor;
 @property(nonatomic,retain) UIColor* fillColor2;
 @property(nonatomic,retain) UIColor* strokeColor;
-@property(nonatomic) int strokeRadius;
+@property(nonatomic) NSInteger strokeRadius;
 @property(nonatomic) UIEdgeInsets backgroundInset;
 
 @end
